GO-6989 Move TestAddToLimitedQueue to batchuploader_test.go
The test covers addToLimitedQueue which is defined in batchuploader.go.
